+++ CHANGELOG.md
@@ -5,6 +5,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.2] - 2021-10-14
+### Added
+- `dependabot.yml`
+- `requirements-splitter.py`
+- `samila_help` function
+- `test.py`
+- `function_test.py`
+- `is_same_data` function
+- `save_image` method
+### Changed
+- `dev-requirements.txt` updated
+- `README.md` updated
+- `__main__.py` updated
+- `test.yml` updated
+- `nft_storage` method updated
 ## [0.1] - 2021-09-30
 ### Added
 - `GenerativeImage` class
@@ -12,7 +27,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `generate` method
 - `nft_storage` method
 
-[Unreleased]: https://github.com/sepandhaghighi/samila/compare/v0.1...dev
+[Unreleased]: https://github.com/sepandhaghighi/samila/compare/v0.2...dev
+[0.2]: https://github.com/sepandhaghighi/samila/compare/v0.1...v0.2
 [0.1]: https://github.com/sepandhaghighi/samila/compare/1058677...v0.1